04_SEARCH/requirements.txt
fastapi
uvicorn
tensorflow  
numpy
pandas
pillow
opencv-python
matplotlib
patool
rarfile
python-dotenv
scipy